requirements.txt
# KUSUMA - RAG System Dependencies
# Production-ready Retrieval-Augmented Generation system

# Core Framework
streamlit>=1.28.0

# Google AI & LLM
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.10
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Data & Computation
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0

# Optional: Vector Store Backends
# Uncomment if using cloud backends
# pinecone-client>=3.0.0
# weaviate-client>=4.0.0
# pymilvus>=2.3.0

# Development Tools (Optional)
# Uncomment for development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# mypy>=1.0.0

pydantic
pypdf
python-docx
tiktok
sentence-transformers
torch